backend/requirements.txt
aiofiles==23.2.1
alembic==1.15.2
annotated-types==0.7.0
anyio==3.7.1
bleach==6.2.0
Brotli==1.1.0
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
cryptography==44.0.2
cssselect2==0.8.0
fastapi==0.104.1
fonttools==4.56.0
greenlet==3.1.1
h11==0.14.0
html2text==2024.2.26
httpcore==1.0.7
httpx==0.28.1
idna==3.10
iniconfig==2.1.0
Jinja2==3.1.2
Mako==1.3.9
MarkupSafe==3.0.2
packaging==24.2
pdfminer.six==20250327
pillow==11.1.0
pluggy==1.5.0
pycparser==2.22
pydantic==2.4.2
pydantic_core==2.10.1
pydyf==0.11.0
pyphen==0.17.2
pytest==8.3.5
pytest-asyncio==0.26.0
pytest-mock==3.14.0
python-multipart==0.0.6
sniffio==1.3.1
SQLAlchemy==2.0.40
starlette==0.27.0
tinycss2==1.4.0
tinyhtml5==2.0.0
typing_extensions==4.13.0
uvicorn==0.23.2
weasyprint==65.0
webencodings==0.5.1
zopfli==0.2.3.post1